db: Refactor detection task in mscoco
This commit will sort the data annotations by the filenames. This is useful for some tasks which require the annotations to be sorted by filename.

Additional fields were introduced to allow the use of the cocoApi by following the same data order as the annotations. This, however, requires the user to use specific fields (like coco_images_ids) to accomplish this.